fraud_detection_fast_api/fraud_detection/requirements.txt
pandas==2.0.3
anthropic==0.3.0
pillow==8.4.0
python-dotenv==1.0.0
pydantic==1.10.2
fastapi
uvicorn
python-multipart
python-dotenv